T0801 - Messages to CI for refunded Birthday Invoices   (#1863)
* T0801 - Messages to CI for refunded Birthday Invoices
 - Fix amount

* T0801 - Messages to CI for refunded Birthday Invoices - Fix :
- Delettrage: deletes the gift if there is no more lettering. If there is remaining lettering, removes the lettering from the gift.
- Reintegration of the possibility to credit after sending the gift.
- If a gift is not sent and it is fully refunded, it is removed.

+++ gift_compassion/models/account_partial_reconcile.py
@@ -25,7 +25,10 @@ class AccountPartialReconcile(models.Model):
         for line in lines.filtered("gift_id"):
             gift = line.gift_id
             if gift.state in ["draft", "verify"]:
-                gift.unlink()
+                if gift.invoice_line_ids:
+                    gift.write({'invoice_line_ids': [(3, line.id)]})
+                    if not gift.invoice_line_ids:
+                        gift.unlink()
             elif gift.state != "Undeliverable":
                 raise UserError(
                     _("You cannot delete the %s. It is already sent to GMC.")
